refactor: split monolithic service traits into focused capability interfaces (Phase 7)
Create focused sub-traits in domain entity modules that the monolithic
service traits now extend:
- OrchestratorControlPlane → WorkflowOrchestrator + ResourceAllocator +
  AgentMonitorService + EventPublisher
- DecisionInbox → DecisionCreator + DecisionResolver
- BoardOrchestrator → IssueDispatcher + IssueApprover

Consumers can now depend on the narrower interface they actually need,
reducing coupling and enabling future module extraction.
